code/requirements.txt
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
fpdf==1.7.2
gast==0.3.3
google-api-core==1.21.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-cloud-vision==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==0.25.3
Pillow==6.2.1
pipenv==2018.11.26
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.4
PyQt5==5.15.0
PyQt5-sip==12.8.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
seaborn==0.9.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-win_amd64.whl
tweepy==3.8.0
twython==3.7.0
urllib3==1.25.9
virtualenv==16.7.7
virtualenv-clone==0.5.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
